Fix failing tests
Story, Product, and Image tests needed to test for the inclusion of
their relationships after those relationships were added.

For the Vendor model, at some point the 'id' field got accidentally
removed from the test. My guess is that it was in a merge conflict --
exactly what these tests were designed to prevent.